Support for existing project.

+++ vagrant/bin/drupal-installation.sh
@@ -4,21 +4,34 @@ set -x
 WORKDIR="/var/www/html"
 VARS_FILE="/drupal/bin/vars.sh"
 
+# Load variables.
+if [ -f "${VARS_FILE}" ]; then
+  source $VARS_FILE
+fi
+
 # Download drupal and libraries.
 if [ ! -f "${WORKDIR}/web/index.php" ]; then
   composer create-project drupal-composer/drupal-project:8.x-dev $WORKDIR --stability dev --no-interaction
 fi
 
-# Install drupal.
-if [ -f "${VARS_FILE}" ] && [ -z "$(${WORKDIR}/vendor/drush/drush/drush status | grep ${MYSQL_CONTAINER})" ]; then
-  source $VARS_FILE
+# Install drush if not already installed.
+if [ ! -f ${WORKDIR}/vendor/drush/drush/drush ] && [ ! -z $DRUSH_VERSION ]; then
+  composer require drush/drush:${DRUSH_VERSION}
+fi
 
+# Make drush accessible globally.
+if [ -f ${WORKDIR}/vendor/drush/drush/drush ]; then
+  ln -s ${WORKDIR}/vendor/drush/drush/drush /usr/local/bin/drush
+fi
+
+# Install drupal.
+if [ -f "${VARS_FILE}" ] && [ -f ${WORKDIR}/vendor/drush/drush/drush ] && [ -z "$(${WORKDIR}/vendor/drush/drush/drush status | grep ${MYSQL_CONTAINER})" ]; then
   # Create symlink for public/private file folder.
   DRUPAL_DEFAULT="${WORKDIR}/web/sites/default"
   MOUNT_DEFAULT=$D_MOUNT_DEFAULT
   for FOLDER in files private
   do
-    if [ -d ${DRUPAL_DEFAULT}/${FOLDER} ]; then
+    if [ ! -L ${DRUPAL_DEFAULT}/${FOLDER} ]; then
       if [ -d ${MOUNT_DEFAULT}/${FOLDER} ]; then rm -rf ${MOUNT_DEFAULT}/${FOLDER}; fi
       mv ${MOUNT_DEFAULT}/${FOLDER} $DRUPAL_DEFAULT
     else
